requirements.txt
# Core dependencies
numpy>=1.26.0,<2.0.0
pandas>=2.0.0
networkx>=3.1
tqdm>=4.65.0

# Machine Learning
torch>=2.0.0
sentence-transformers>=2.7.0
transformers>=4.30.0
faiss-cpu>=1.7.4

# API clients
openai>=1.0.0
tenacity>=9.1.2

# Data processing
datasets>=2.14.0
jsonschema>=4.17.3

# Configuration
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.2

